nPencils as vector<uint> and correctly initialised.

+++ vlasovsolver/cpu_trans_map_amr.hpp
@@ -202,7 +202,7 @@ setOfPencils buildPencilsWithNeighbors( const dccrg::Dccrg<SpatialCell,dccrg::Ca
 bool trans_map_1d_amr(const dccrg::Dccrg<spatial_cell::SpatialCell,dccrg::Cartesian_Geometry>& mpiGrid,
                   const std::vector<CellID>& localPropagatedCells,
                   const std::vector<CellID>& remoteTargetCells,
-                  std::vector<uint64_t>& nPencils,
+                  std::vector<uint>& nPencils,
                   const uint dimension,
                   const Realv dt,
                   const uint popID);
